fixed queue and random problem

+++ include/Event.h
@@ -3,17 +3,17 @@
 
 class Event{
 public:
-    Event(int _type, int _id, float _time); //for Call INITIATION
-    Event(int _type, int _id, float _time, int _baseStation, bool _dir, float _speed, float _duration); //for Call HANDOVER
-    Event(int _type, int _id, float _time, int _baseStation); //for Call TERMINATION
+    Event(int _type, int _id, double _time); //for Call INITIATION
+    Event(int _type, int _id, double _time, int _baseStation, bool _dir, double _speed, double _duration); //for Call HANDOVER
+    Event(int _type, int _id, double _time, int _baseStation); //for Call TERMINATION
     ~Event(){};
     int type; //1 for initiation, 2 for handover, 3 for termination
     int id;
-    float time; //when it happens
+    double time; //when it happens
     int baseStation;
     bool dir;
-    float speed;
-    float duration;
+    double speed;
+    double duration;
     bool operator <(const Event& e);
 };
